QUANTUM-6 Add test for quantum.js secondDerivative.
This is an aboslutely awful function. It takes in the previous term and
the next. One fun fact, is that since the accuracy of Math.cos is only
so much, you rapidly get into the situation where adding additional
zeros to dx makes the approximation less accurate.

Fun with floating point! Huzzah!