QUANTUM-6 Fixing state generation, adding more tests.
State generation was being handled improperly. Now we succesfully
generate states for each particle. I believe the intention is to use
the box geometry to map a given state to physical reality.

That is, the state [2,1] (of a box of resolution 10, [0,1]) would be
passed into geometry.resolve(state=[2,1]) and would, for a box, return
{"particle1" : {"x": 2/10}, "particle2" : {"x" : 1/10}}.

Granted, this is assuming that the states are in a position-like basis,
I'll need to think more about how I want this to work.