QUANTUM-6 Making sense of StateFunction.
StateFunction is the WaveFunction constructor. As such, it takes a list
of basis states, and their associated amplitudes. The Hamiltonian will
need to take in a list of basis states, and a way of computing the
projection onto those. Lots of comments being written out; more work
to be done attempting to construct how this is all going to work.

Suffice to say, it looks like when I wrote this I was thinking about
the WaveFunction as it's own thing, rather than just a point in Hilbert
Space. This is forcing me to do all sorts of strange things. Hopefully,
the final output will be sensical. I'm probably going to need to
rethink all of this.