QUANTUM-XXX Expand on the plan here.
More detailed explanation of my concerns with typical presentations of
Quantum Mechnics (at least based on impressions at the time of taking
said courses) and an explanation of the proposed plan of attack.

This hopefully explains why combinatorics, System, and statefunction
exist. We're trying to emphasize the Linear Algebra aspects here, which
is the "Wave Function is the probability of being in a particular
System Configuration State".